+++ ls-notes/src/Components/CreateNoteView.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+class CreateNoteView extends React.Component {
+  render() {
+    return <h1>I'm CreateNoteView{this.props.title}</h1>;
+  }
+}
