+++ ls-notes/src/Components/DeleteNoteModal.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+class DeleteNoteModal extends React.Component {
+  render() {
+    return <h1>I'm DeleteNoteModal{this.props.title}</h1>;
+  }
+}
+
