+++ ls-notes/src/Components/EditNoteView.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+class EditNoteView extends React.Component {
+  render() {
+    return <h1>I'm EditNoteView{this.props.title}</h1>;
+  }
+}
