+++ ls-notes/src/Components/NoteDetailView.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+class NoteDetailView extends React.Component {
+  render() {
+    return <h1>I'm {this.props.title}</h1>;
+  }
+}
