+++ ls-notes/src/Components/NoteListView.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+class NoteListView extends React.Component {
+  render() {
+    return (<h1>I'm NoteListView {this.props.title}</h1>
+  );
+  }
+}
